refactor: 트랜잭션 범위 재조정으로 커넥션 풀 고갈 문제 개선
