+++ src/main/java/org/umc/valuedi/domain/asset/service/command/AssetSyncProcessor.java
@@ -4,7 +4,6 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.scheduling.annotation.Async;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 import org.umc.valuedi.domain.asset.dto.res.AssetResDTO;
 import org.umc.valuedi.domain.connection.service.command.SyncLogCommandService;
 import org.umc.valuedi.domain.ledger.service.command.LedgerSyncService;
@@ -29,7 +28,6 @@ public class AssetSyncProcessor {
      * 실제 동기화 로직을 수행하는 비동기 메서드
      */
     @Async("assetFetchExecutor")
-    @Transactional
     public void runSyncProcess(Long memberId, Long logId) {
         log.info("자산 동기화 백그라운드 작업을 시작합니다. 회원 ID: {}", memberId);
         try {
@@ -49,9 +47,9 @@ public class AssetSyncProcessor {
             }
 
             // 트랜잭션 3: 동기화 로그 및 최종 시간 업데이트
-            member.updateLastSyncedAt();
+            ledgerSyncService.updateMemberLastSyncedAt(memberId);
             syncLogCommandService.updateToSuccess(logId);
-            log.info("자산 동기화 백그라운드 작업을 성공적으로 완료했습니다. 회원 ID: {}", member.getId());
+            log.info("자산 동기화 백그라운드 작업을 성공적으로 완료했습니다. 회원 ID: {}", memberId);
 
         } catch (Exception e) {
             // 실패 로그 기록
