+++ src/main/resources/application.yml
@@ -35,6 +35,10 @@ spring:
           connectiontimeout: 5000
           timeout: 5000
           writetimeout: 5000
+  hikari:
+    maximum-pool-size: 30
+    leak-detection-threshold: 5000 # 커넥션 점유 메서드 파악
+
 codef:
   client-id: ${CODEF_CLIENT_ID}
   client-secret: ${CODEF_CLIENT_SECRET}
